linted

+++ test/JsonWriter.ints.t.sol
@@ -8,22 +8,22 @@ contract JsonWriterIntTest is Test {
     using JsonWriter for JsonWriter.Json;
 
     struct IntTestCase {
-        int arg;
+        int256 arg;
         string expected;
     }
 
     function fixtureValues() public pure returns (IntTestCase[] memory) {
         IntTestCase[] memory entries = new IntTestCase[](5);
-        entries[0] = IntTestCase(-1, '-1');
-        entries[1] = IntTestCase(0, '0');
-        entries[2] = IntTestCase(1, '1');
+        entries[0] = IntTestCase(-1, "-1");
+        entries[1] = IntTestCase(0, "0");
+        entries[2] = IntTestCase(1, "1");
         entries[3] = IntTestCase(
-            -57896044618658097711785492504343953926634992332820282019728792003956564819968, 
-            '-57896044618658097711785492504343953926634992332820282019728792003956564819968'
+            -57896044618658097711785492504343953926634992332820282019728792003956564819968,
+            "-57896044618658097711785492504343953926634992332820282019728792003956564819968"
         );
         entries[4] = IntTestCase(
-            57896044618658097711785492504343953926634992332820282019728792003956564819967, 
-            '57896044618658097711785492504343953926634992332820282019728792003956564819967'
+            57896044618658097711785492504343953926634992332820282019728792003956564819967,
+            "57896044618658097711785492504343953926634992332820282019728792003956564819967"
         );
 
         return entries;
@@ -42,7 +42,7 @@ contract JsonWriterIntTest is Test {
         json = json.writeIntValue(-1);
         json = json.writeEndArray();
 
-        assertEq(json.value, '[-1]');
+        assertEq(json.value, "[-1]");
     }
 
     function test_writesArrayWithMultipleIntValues() public pure {
@@ -52,7 +52,7 @@ contract JsonWriterIntTest is Test {
         json = json.writeIntValue(1);
         json = json.writeEndArray();
 
-        assertEq(json.value, '[-1,1]');
+        assertEq(json.value, "[-1,1]");
     }
 
     function fixtureProperties() public pure returns (IntTestCase[] memory) {
@@ -60,15 +60,21 @@ contract JsonWriterIntTest is Test {
         entries[0] = IntTestCase(-1, '"prop": -1');
         entries[1] = IntTestCase(0, '"prop": 0');
         entries[2] = IntTestCase(1, '"prop": 1');
-        entries[3] = IntTestCase(-57896044618658097711785492504343953926634992332820282019728792003956564819968, '"prop": -57896044618658097711785492504343953926634992332820282019728792003956564819968');
-        entries[4] = IntTestCase(57896044618658097711785492504343953926634992332820282019728792003956564819967, '"prop": 57896044618658097711785492504343953926634992332820282019728792003956564819967');
+        entries[3] = IntTestCase(
+            -57896044618658097711785492504343953926634992332820282019728792003956564819968,
+            '"prop": -57896044618658097711785492504343953926634992332820282019728792003956564819968'
+        );
+        entries[4] = IntTestCase(
+            57896044618658097711785492504343953926634992332820282019728792003956564819967,
+            '"prop": 57896044618658097711785492504343953926634992332820282019728792003956564819967'
+        );
 
         return entries;
     }
 
     function table_writesIntPropertyOf(IntTestCase memory properties) public pure {
         JsonWriter.Json memory json;
-        json = json.writeIntProperty('prop', properties.arg);
+        json = json.writeIntProperty("prop", properties.arg);
 
         assertEq(json.value, properties.expected);
     }
@@ -76,7 +82,7 @@ contract JsonWriterIntTest is Test {
     function test_writesObjectWithSingleIntPropertyAndValue() public pure {
         JsonWriter.Json memory json;
         json = json.writeStartObject();
-        json = json.writeIntProperty('prop', -1);
+        json = json.writeIntProperty("prop", -1);
         json = json.writeEndObject();
 
         assertEq(json.value, '{"prop": -1}');
@@ -85,8 +91,8 @@ contract JsonWriterIntTest is Test {
     function test_writesObjectWithMultipleIntPropertiesAndValues() public pure {
         JsonWriter.Json memory json;
         json = json.writeStartObject();
-        json = json.writeIntProperty('prop1', -1);
-        json = json.writeIntProperty('prop2', 1);
+        json = json.writeIntProperty("prop1", -1);
+        json = json.writeIntProperty("prop2", 1);
         json = json.writeEndObject();
 
         assertEq(json.value, '{"prop1": -1,"prop2": 1}');
